Multi GPU test to incremental_dump of ScoredHashTable
