+++ corelib/dynamicemb/test/unit_tests/table_operation/test_table_operation.py
@@ -22,13 +22,7 @@ import torch
 import torch.distributed as dist
 import torchrec
 from dynamicemb.scored_hashtable import ScoreArg, ScoreSpec, get_scored_table
-from dynamicemb_extensions import (
-    InsertResult,
-    ScorePolicy,
-    device_timestamp,
-    table_count_matched,
-    table_partition,
-)
+from dynamicemb_extensions import InsertResult, ScorePolicy, table_partition
 
 
 @pytest.fixture
@@ -597,174 +591,3 @@ def test_table_reserve(
             print(
                 f"Table reserve passed from capacity({table.capacity() // 2}) to capacity({table.capacity()})"
             )
-
-
-def table_num_matched(table, threshold) -> int:
-    d_num_matched = torch.zeros(1, dtype=torch.int64, device=table.device)
-    table_count_matched(
-        table.table_storage_,
-        table.fileds_type_,
-        table.bucket_capacity_,
-        [threshold],
-        d_num_matched,
-    )
-    return d_num_matched.cpu().item()
-
-
-@pytest.mark.parametrize(
-    "bucket_capacity, batch_size, num_buckets, num_iteration, dump_interval",
-    [
-        pytest.param(
-            128, 128, 4096, 8192, 1024, id="Never evict keys from current batch"
-        ),
-        pytest.param(128, 65536, 4096, 32, 8),
-        pytest.param(
-            128, 1024 + 13, 8, 12, 3, id="Always evict keys from current batch"
-        ),
-        pytest.param(
-            128, 1024, 32, 32, 4, id="Always evict keys from last dump_interval"
-        ),
-        pytest.param(512, 512, 1024, 2048, 256, id="Different bucket capacity"),
-    ],
-)
-@pytest.mark.parametrize(
-    "score_policy",
-    [ScorePolicy.ASSIGN, ScorePolicy.ACCUMULATE, ScorePolicy.GLOBAL_TIMER],
-)
-def test_table_incremental_dump(
-    request,
-    bucket_capacity,
-    batch_size,
-    num_buckets,
-    num_iteration,
-    dump_interval,
-    score_policy,
-):
-    print(f"\n{request.node.name}")
-
-    table = get_scored_table(
-        capacity=num_buckets * bucket_capacity,
-        bucket_capacity=bucket_capacity,
-        key_type=torch.int64,
-        score_specs=[ScoreSpec(name="score1", policy=score_policy)],
-    )
-
-    assert torch.cuda.is_available()
-    device = torch.cuda.current_device()
-
-    # insert once: the first insert will insert all keys, and no eviction(batch = bucket_capacity)
-    keys = torch.randperm(bucket_capacity, device=device, dtype=torch.int64)
-    score_args = [
-        ScoreArg(name="score1", value=get_scores(score_policy, keys), is_return=True)
-    ]
-    insert_results = torch.empty(
-        bucket_capacity, dtype=table.result_type, device=device
-    ).fill_(InsertResult.INIT.value)
-
-    table.insert(keys, score_args, None, insert_results)
-
-    # check 1: count_matched works well
-    assert table_num_matched(table, 0) == bucket_capacity
-
-    # check 2: table.incremental_dump is consistent with count_matched
-    dumped_keys, dumped_named_scores = table.incremental_dump({"score1": 0})
-    assert dumped_keys.numel() == bucket_capacity
-    assert set(dumped_keys.tolist()) == set(keys.tolist())
-    assert (insert_results == InsertResult.INSERT.value).sum() == bucket_capacity
-
-    # insert iteratively
-    offset = bucket_capacity
-
-    for i in range(0, num_iteration, dump_interval):
-        if score_policy == ScorePolicy.GLOBAL_TIMER:
-            undumped_score = device_timestamp()
-        elif score_policy == ScorePolicy.ASSIGN:
-            global score_step
-            undumped_score = score_step + 1
-        else:
-            undumped_score = 1
-
-        # pre-check: using uninsert score will dump nothing.
-        if score_policy != ScorePolicy.ACCUMULATE:
-            dumped_keys, _ = table.incremental_dump({"score1": undumped_score})
-            assert dumped_keys.numel() == 0
-
-        interval_keys = torch.arange(
-            offset,
-            offset + dump_interval * batch_size,
-            device=device,
-            dtype=torch.int64,
-        )
-        interval_existed_in_table = torch.empty(
-            dump_interval * batch_size, device=device, dtype=torch.bool
-        ).fill_(True)
-        interval_offset = offset
-
-        num_remain = dump_interval * batch_size
-
-        for j in range(dump_interval):
-            keys = torch.randperm(batch_size, device=device, dtype=torch.int64) + offset
-            offset += batch_size
-            score_args[0].value = get_scores(score_policy, keys)
-            insert_results = torch.empty(
-                batch_size, dtype=table.result_type, device=device
-            ).fill_(InsertResult.INIT.value)
-
-            old_size = table.size()
-            num_evict, evict_keys, _, evict_named_scores = table.insert_and_evict(
-                keys, score_args, None, insert_results
-            )
-            new_size = table.size()
-
-            num_inserted = (insert_results == InsertResult.INSERT.value).sum()
-            num_reclaim = (insert_results == InsertResult.RECLAIM.value).sum()
-            num_assign = (insert_results == InsertResult.ASSIGN.value).sum()
-            num_inserted_by_eviction = (
-                insert_results == InsertResult.EVICT.value
-            ).sum()
-            num_insert_failed = (insert_results == InsertResult.BUSY.value).sum()
-
-            assert num_assign == 0
-            assert num_reclaim == 0
-
-            assert new_size - old_size == num_inserted
-            assert num_inserted_by_eviction + num_insert_failed == num_evict
-            assert (
-                num_inserted + num_inserted_by_eviction + num_insert_failed
-                == batch_size
-            )
-
-            # evicted keys are not in the table anymore, including insert failed keys.
-            evict_keys_interval_mask = (evict_keys - interval_offset) >= 0
-            evict_keys_interval_mask = (
-                evict_keys[evict_keys_interval_mask] - interval_offset
-            )
-            interval_existed_in_table[evict_keys_interval_mask] = False
-
-            num_remain -= evict_keys_interval_mask.numel()
-
-        # check 3: incremental dump as expected
-        keys, named_scores = table.incremental_dump({"score1": undumped_score})
-        if score_policy != ScorePolicy.ACCUMULATE:
-            assert ((named_scores["score1"].to(device) - undumped_score) >= 0).all()
-            assert keys.numel() == num_remain
-            ascend_dumped_keys, _ = torch.sort(keys.to(device))
-            ascend_interval_keys, _ = torch.sort(
-                interval_keys[interval_existed_in_table]
-            )
-            assert torch.equal(ascend_dumped_keys, ascend_interval_keys)
-        else:
-            assert keys.numel() == table.size()
-            assert torch.isin(
-                interval_keys[interval_existed_in_table], keys.to(device)
-            ).all()
-
-        # check 4: using min_score to count will get table's size
-        dumped_keys, _ = table.incremental_dump({"score1": 0})
-        assert dumped_keys.numel() == table.size()
-
-        # log
-        load_factor = table.size() / table.capacity()
-        print(
-            f"Load factor={load_factor:.3f}, there are {interval_existed_in_table.sum()} keys existed in the table for the last interval {dump_interval*batch_size}"
-        )
